Removes legacy schema class configuration
Replaces the old Java-based configuration with a configuration file approach, enhancing flexibility and eliminating the need for recompilation.

Also, updates the README and CLAUDE documentation to reflect the change to property file configuration.
