--- src/main/java/Simulator/Parameters.java
@@ -1,10 +1,0 @@
-package Simulator;
-
-public class Parameters
-{
-    /**
-     * The delay between replication time in dynamic replication when there exists more than a
-     * single data owner.
-     */
-    public static final int REPLICATION_TIME_INTERVAL = 1;
-}
